pyproject.toml
[project]
name = "modrinthbot"
version = "0.0.1"
description = "An official utility discord bot for Modrinth"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
]
